# src/backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Task Management API",
      "version": "1.0.0",
      "description": "A simple API for task management"
    },
    "servers": [
      {
        "url": "http://localhost:5000"
      }
    ],
    "paths": {
      "/tasks": {
        "get": {
          "summary": "Get all tasks",
          "operationId": "getTasks",
          "responses": {
            "200": {
              "description": "A list of tasks",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new task",
          "operationId": "createTask",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Task created successfully"
            }
          }
        }
      },
      "/tasks/{taskId}": {
        "get": {
          "summary": "Get a task by ID",
          "operationId": "getTaskById",
          "parameters": [
            {
              "name": "taskId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Task found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            },
            "404": {
              "description": "Task not found"
            }
          }
        },
        "patch": {
          "summary": "Update a task",
          "operationId": "updateTask",
          "parameters": [
            {
              "name": "taskId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Task updated"
            },
            "404": {
              "description": "Task not found"
            }
          }
        },
        "delete": {
          "summary": "Delete a task",
          "operationId": "deleteTask",
          "parameters": [
            {
              "name": "taskId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Task deleted"
            },
            "404": {
              "description": "Task not found"
            }
          }
        }
      },
      "/users": {
        "get": {
          "summary": "Get all users",
          "operationId": "getUsers",
          "responses": {
            "200": {
              "description": "A list of users",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new user",
          "operationId": "createUser",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "User created successfully"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Task": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "title": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "owner": {
              "type": "string"
            },
            "status": {
              "type": "string",
              "enum": ["Pending", "In Progress", "Completed"]
            },
            "deadline": {
              "type": "string",
              "format": "date"
            }
          }
        },
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "surName": {
              "type": "string"
            },
            "fatherName": {
              "type": "string"
            },
            "departmentName": {
              "type": "string"
            },
            "position": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "role": {
              "type": "string",
              "enum": ["user", "admin"]
            },
            "isDeleted": {
              "type": "boolean"
            }
          }
        }
      }
    }
  }
  